Force glibc 2.17 compat, prevent *.lib in windows wheels

+++ hatch_build.py
@@ -7,6 +7,8 @@ from hatchling.builders.hooks.plugin.interface import BuildHookInterface
 from packaging import tags
 import os
 
+library_suffixes = ['.so', '.dll', '.dylib']
+
 cpus = [
     {'cpu': 'x86_64'}, 
     {'cpu': 'haswell', 'opt_level': 'v3'},
@@ -99,7 +101,7 @@ class CustomHook(BuildHookInterface[Any]):
                     subprocess.run(["python-zig", "build", "-Doptimize=ReleaseFast", f"-Dtarget={zig_target}", f"-Dcpu={cpu_spec['cpu']}"], check=True)
 
                     for file_path in self.source_dir.rglob("*"):
-                        if file_path.is_file():
+                        if file_path.is_file() and file_path.suffix in library_suffixes:
                             if 'opt_level' in cpu_spec:
                                 name = file_path.stem + f".{cpu_spec['opt_level']}" + file_path.suffix
                                 shutil.copy2(file_path, Path(self.target_dir, name))
@@ -129,7 +131,7 @@ class CustomHook(BuildHookInterface[Any]):
         
             # Copy the compiled binaries
             for file_path in self.source_dir.rglob("*"):
-                if file_path.is_file():
+                if file_path.is_file() and file_path.suffix in library_suffixes:
                     shutil.copy2(file_path, self.target_dir)
